feat: Finish particles system - Add ParticleTransform class to handle particle transform - Fix wrong passCB set in ShaderParticle - Fix wrong InstanceData manipulation when copying data TODO: - Check why so few particles appear on screen - Add ParticleSystem class to remove particle update from ParticleRenderer

+++ ChokbarEngine/Source/Engine/GameTimer.h
@@ -15,17 +15,17 @@ namespace Chokbar
 		void Stop();	// Call when paused.
 		void Tick();	// Call every frame.
 
-		float GetCurrentFrameTime();
+		double GetCurrentFrameTime();
 
 		float GetTotalTime();
 		float GetGameTime() const { return m_CurrTime; }
-		static float GetDeltaTime() { return m_DeltaTime; }
+		static double GetDeltaTime() { return m_DeltaTime; }
 
 
 	private:
 
 		double m_SecondsPerCount;
-		static float m_DeltaTime;
+		static double m_DeltaTime;
 
 		float m_BaseTime;
 		float m_PausedTime;
